add file to save history

+++ src/main/java/forecast/by/month/service/ExcelWriter.java
@@ -2,10 +2,11 @@ package forecast.by.month.service;
 
 import org.apache.poi.ss.usermodel.Workbook;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ExcelWriter {
     Workbook createWorkbookWithSheets(List<String> monthNames);
-    Double getTotalServiceTeam(Workbook inputWorkbook, String serviceTeam, String sheetName);
+    BigDecimal getTotalServiceTeam(Workbook inputWorkbook, String serviceTeam, String sheetName);
     void copyServiceHoursSheetData(Workbook inputWorkbook, Workbook outputWorkbook, String serviceTeam, String HorasServicioSheetName, String ServiceHoursSheetName, String ajustesSheetName, String facturacionSheetName);
 }
